ML API/requirements.txt
fastapi
uvicorn
tensorflow
pillow
